Cleanup DLP Quickstart.
